Implement bandwidth support
This lets some interfaces be "faster" than others and send more packets
at a time.

Had to also add in "forwarding capacity" support on the switch so it
wasn't just doing one packet per port per step.

And had to rework the queue implementation on the port (using lists
instead of the Queue class)