Update ARP table from regular IP packets
Updates the ARP table from regular IP (non-ARP) packets received. This
way, if we receive a normal packet, we don't have to waste time with an
ARP request to respond.

This also adds filtering to not add things to the ARP table if they're
not in our network(s) or are on the wrong port for that network.