Fix application logging
Kinda multi-part. Added "step tracking" to applications, so it knows how
many steps since it was started. Added more details to the logs for DHCP
Server and Client so you can tell the different requests apart. Added
HWID src/dst for packet processing, so that it's avaialble to
applications... if they want it (and good for logging in DHCP). And then
made the log window bigger by default.